use RPATH by default
Add option `CUDA_MEMTEST__ADD_RPATH` to control if cuda_memtest is setting the rpath.
The new default is that RPATH will be used but optionally allow disabling this and falling back to the old behaviori.

Using rpath's is motivated by the spack where `LD_LIBRARY_PATH` is not set anymore: spack/spack#28354